+++ examples/complete/locals.tf
@@ -1,4 +1,3 @@
 locals {
   email = "example.email@example.com"
-  name  = "complete_example_budget"
 }