+++ main.tf
@@ -12,14 +12,12 @@ resource "aws_budgets_budget" "main" {
   dynamic "auto_adjust_data" {
     for_each = length(keys(var.auto_adjust_data)) > 0 ? [var.auto_adjust_data] : []
     content {
-      auto_adjust_type      = lookup(var.auto_adjust_data, "auto_adjust_type", null)
-      last_auto_adjust_time = lookup(var.auto_adjust_data, "last_auto_adjust_time", null)
+      auto_adjust_type = lookup(var.auto_adjust_data, "auto_adjust_type", null)
 
       dynamic "historical_options" {
         for_each = try([var.auto_adjust_data["historical_options"]], [])
         content {
-          budget_adjustment_period   = try(historical_options.value.budget_adjustment_period, null)
-          lookback_available_periods = try(historical_options.value.lookback_available_periods, null)
+          budget_adjustment_period = try(historical_options.value.budget_adjustment_period, null)
         }
       }
     }